refactor(app): extract module-level helpers to runtime/ (#69)
Move the seven channel adapters, seven builder functions, and two demo
helpers out of the 5 344-line app.py into a new src/vanchor/runtime/
sub-package (channels.py, builders.py, demo.py + __init__.py).
app.py re-exports every name so all existing imports — including the
two test files that import _TeeMotor directly from vanchor.app — keep
working without change.  Pure behaviour-preserving move; no logic
altered.
